backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
Markdown==3.9
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
wheel==0.45.1
